src/Python/Scripts/GTK/dark-overlay/requirements.txt
EasyProcess==0.2.7
Pillow==6.1.0
pycairo==1.18.1
PyGObject==3.34.0
pyscreenshot==0.5.1
setproctitle==1.1.10
